attempt to fix commander pt2 (kind of a refactor)
