+++ src/commands/export.ts
@@ -1,9 +1,11 @@
+import * as DenoJson from "../../deno.json" with { type: "json" };
 import type { TheExporterConstructedParams } from "./constructors/command.ts";
 import { GetProjectEnvironment, SpotProject } from "../functions/projects.ts";
 import { JoinPaths } from "../functions/filesystem.ts";
 import { ColorString, LogStuff, StringifyYaml } from "../functions/io.ts";
-import { APP_NAME, FULL_NAME, VERSION } from "../constants.ts";
+import { FULL_NAME } from "../constants.ts";
 import { GetDateNow } from "../functions/date.ts";
+import { APP_NAME } from "../constants/name.ts";
 
 export default function TheExporter(params: TheExporterConstructedParams) {
     const { project } = params;
@@ -16,7 +18,7 @@ export default function TheExporter(params: TheExporterConstructedParams) {
     const outFileName = params.json === true ? "fknode-cpf.jsonc" : "fknode-cpf.yaml";
 
     const comment =
-        `# Generated by ${FULL_NAME}\n# ${APP_NAME.CASED} v${VERSION}\n# This has been manually generated at ${GetDateNow()}. Content here isn't synced with your project.`;
+        `# Generated by ${FULL_NAME}\n# ${APP_NAME.CASED} v${DenoJson.default.version}\n# This has been manually generated at ${GetDateNow()}. Content here isn't synced with your project.`;
 
     const commentString = params.json === true ? comment.replaceAll("#", "//") : comment;
