+++ src/constants/fwords.ts
@@ -1,0 +1,48 @@
+/**
+ * Different variants of the f-word for in-app usage. Not fully "explicit" as an asterisk is used, like in f*ck.
+ */
+export const FWORDS: {
+    /**
+     * Base word. 4 letters.
+     *
+     * @type {string}
+     */
+    FK: string;
+    /**
+     * Base word but with -ing.
+     *
+     * @type {string}
+     */
+    FKN: string;
+    /**
+     * Noun. What we call a project that's made with NodeJS. Base word but mentioning his mother (-er).
+     *
+     * @type {string}
+     */
+    MF: string;
+    /**
+     * Plural for `mf`.
+     *
+     * @type {string}
+     */
+    MFS: string;
+    /**
+     * Adjective. What we describe a project that's made with NodeJS as.
+     *
+     * @type {string}
+     */
+    MFN: string;
+    /**
+     * _"Something went **mother** + `fkn` + **ly**"_
+     *
+     * @type {string}
+     */
+    MFLY: string;
+} = {
+    FK: "f*ck",
+    FKN: "f*cking",
+    MF: "m*therf*cker",
+    MFS: "m*therf*ckers",
+    MFN: "m*therf*cking",
+    MFLY: "m*therf*ckingly",
+};
