+++ src/constants/name.ts
@@ -1,0 +1,10 @@
+/**
+ * Best CLI app ever (it's name, so you don't, for example, miss-capitalize it).
+ *
+ * @type {{CASED: string, CLI: string, SCOPE: string}}
+ */
+export const APP_NAME: { CASED: string; CLI: string; SCOPE: string } = {
+    CASED: "FuckingNode",
+    CLI: "fuckingnode",
+    SCOPE: "@zakahacecosas/fuckingnode",
+};
