+++ src/functions/config.ts
@@ -1,4 +1,4 @@
-import { APP_NAME, DEFAULT_SCHEDULE_FILE, DEFAULT_SETTINGS, FWORDS, LOCAL_PLATFORM } from "../constants.ts";
+import { DEFAULT_SCHEDULE_FILE, DEFAULT_SETTINGS } from "../constants.ts";
 import type { CF_FKNODE_SETTINGS } from "../types/config_files.ts";
 import { FknError } from "./error.ts";
 import { BulkRemove, CheckForPath, JoinPaths } from "./filesystem.ts";
@@ -6,6 +6,9 @@ import { parse as parseYaml } from "@std/yaml";
 import { ColorString, Interrogate, LogStuff, StringifyYaml } from "./io.ts";
 import { type UnknownString, validate, validateAgainst } from "@zakahacecosas/string-utils";
 import { format } from "@std/fmt/bytes";
+import { FWORDS } from "../constants/fwords.ts";
+import { LOCAL_PLATFORM } from "../constants/platform.ts";
+import { APP_NAME } from "../constants/name.ts";
 
 /**
  * Returns file paths for all config files the app uses.
